MueLu: kokkos refactor of phase 3 and 2a
Mainly wrapping phase 3 in a parallel_for loop using a Kokkos range policy
with a single index to run on serially on device.
Also cleaning up a view initialization in Phase 2a to make sure that the
initialization happens in host mirror view and not device view (would fail
with CUDA otherwise).